release-330
= 3.3.0 =
- new option to use an external host for Shariff and the share count
backend
- new share count service OpenShareCount.com for Twitter
- new settings tab "Statistic" for all options regarding the share
counts
- new settings tab "Status" for all system checks to increase
performance
- new design preview button without Twitter
- fix counter VK on design round
- fix Facebook total_count again
- fix double Twitter share windows under certain conditions
- reactivate Flattr counts, since they fixed their API
- purge Shariff transients on update, deactivate and uninstall
- code cleanup
- add vk to help section
- add a know bugs section to the readme
