+++ shariff/locale/shariff3UU-de_DE.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: Shariff Wrapper for Wordpress\n"
-"PO-Revision-Date: 2015-11-20 01:27+0100\n"
+"PO-Revision-Date: 2015-11-26 02:01+0100\n"
 "Last-Translator: JP\n"
 "Plural-Forms: nplurals=2; plural=n != 1;\n"
 "Content-Type: text/plain; charset=utf-8\n"
@@ -459,8 +459,8 @@ msgstr "Verlange die Eingabe einer Absenderadresse."
 # @ shariff3UU
 #. Text in function
 #: shariff.php:1
-msgid "Select the desired services in the order you want them to be displayed, where the Shariff buttons should be included automatically and if you want the share counts to be shown."
-msgstr "Wähle die zu aktivierenden Dienste in der gewünschten Reihenfolge, wo die Shariff-Buttons automatisch eingebunden werden sollen und ob die Share-Statistik in den Buttons angezeigt werden soll."
+msgid "Select the desired services in the order you want them to be displayed and where the Shariff buttons should be included automatically."
+msgstr "Wähle die zu aktivierenden Dienste in der gewünschten Reihenfolge und wo die Shariff-Buttons automatisch eingebunden werden sollen."
 
 # @ shariff3UU
 #. Text in function
@@ -821,3 +821,45 @@ msgstr "Textauszug (Excerpt)"
 #: shariff.php:1
 msgid "Disable the following services (share counts only):"
 msgstr "Deaktiviere die folgenden Dienste (nur Statistik):"
+
+# @ shariff3UU
+#. Text in function
+#: shariff.php:1
+msgid "Shariff Wrapper: Twitter has disabled share counts. Please read <a href=\"https://www.jplambeck.de/twitter-saveoursharecounts/\" target=\"_blank\"><strong>this post</strong></a> prior to activating the alternative via OpenShareCount.com!"
+msgstr "Shariff Wrapper: Twitter hat die share counts abgeschaltet. Bitte vor Aktivierung der Alternative über OpenShareCount.com <a href=\"https://www.jplambeck.de/twitter-saveoursharecounts/\" target=\"_blank\"><strong>diesen Beitrag</strong></a> lesen!"
+
+# @ shariff3UU
+#. Text in function
+#: shariff.php:1
+msgid "Statistic"
+msgstr "Statistik"
+
+# @ shariff3UU
+#. Text in function
+#: shariff.php:1
+msgid "This determines how share counts are handled by Shariff."
+msgstr "Diese Einstellungen bestimmen, wie Shariff mit den sogenannten „Share Counts“ umgeht."
+
+# @ shariff3UU
+#. Text in function
+#: shariff.php:1
+msgid "External host for share counts, shariff.js and CSS:"
+msgstr "Externer Server für die Share Counts, shariff.js und CSS:"
+
+# @ shariff3UU
+#. Text in function
+#: shariff.php:1
+msgid "Warning: This is an <strong>experimental</strong> feature. Please read the <a href=\"https://wordpress.org/plugins/shariff/faq/\" target=\"_blank\">Frequently Asked Questions (FAQ)</a>."
+msgstr "Warnung: Dies ist eine <strong>experimentelle</strong> Funktion. Bitte lies die <a href=\"https://wordpress.org/plugins/shariff/faq/\" target=\"_blank\">Frequently Asked Questions (FAQ)</a>."
+
+# @ shariff3UU
+#. Text in function
+#: shariff.php:1
+msgid "Please check, if you have to add this domain to the array $SHARIFF_FRONTENDS on the external server."
+msgstr "Bitte überprüfe, ob diese Domain zum Array $SHARIFF_FRONTENDS auf dem externen Server hinzugefügt werden muss."
+
+# @ shariff3UU
+#. Text in function
+#: shariff.php:1
+msgid " <span style=\"color: red; font-weight: bold;\">Warning:</span> You entered an external host! Therefore most options on this page have no effect. You need to configure them on the external server. Remember: This feature is still experimental!"
+msgstr " <span style=\"color: red; font-weight: bold;\">Warnung:</span> Es wurde ein externer Server angegeben! Daher haben viele Optionen auf dieser Seite keinen Effekt. Diese müssen auf dem externen Server eingestellt werden. Zur Erinnerung: Diese Funktion ist noch experimentell!"
